+++ R/change_analysis.R
@@ -36,22 +36,22 @@
 #' Estimation of change between two probability surveys
 #'
 #' This function organizes input and output for estimation of change between two
-#' probability surveys.  The input data argument can be either a data frame or a
-#' simple features (\code{sf}) object.  If an \code{sf} object is used,
-#' coordinates are extracted from the geometry column in the object, arguments
-#' \code{xcoord} and \code{ycoord} are assigned values \code{"xcoord"} and
-#' \code{"ycoord"}, respectively, and the geometry column is dropped from the
-#' object.
+#' probability surveys.  The analysis data,
+#' \code{dframe}, can be either a data frame or a simple features (\code{sf}) object.  If an
+#' \code{sf} object is used, coordinates are extracted from the geometry column in the
+#' object, arguments \code{xcoord} and \code{ycoord} are assigned values
+#' \code{"xcoord"} and \code{"ycoord"}, respectively, and the geometry column is
+#' dropped from the object.
 #'
-#' @param dframe Data frame containing survey design variables, response
+#' @param dframe The analysis data. A data frame containing survey design variables, response
 #'   variables, and subpopulation (domain) variables.
 #'
 #' @param vars_cat Vector composed of character values that identify the
-#'   names of categorical response variables in the dframe data frame.  The
+#'   names of categorical response variables in \code{dframe}.  The
 #'   default is \code{NULL}.
 #'
 #' @param vars_cont Vector composed of character values that identify the
-#'   names of continuous response variables in the dframe data frame.  The
+#'   names of continuous response variables in \code{dframe}.  The
 #'   default is \code{NULL}.
 #'
 #' @param test Character string or character vector providing the location
@@ -60,14 +60,14 @@
 #'   The default is \code{"mean"}.
 #'
 #' @param subpops Vector composed of character values that identify the
-#'   names of subpopulation (domain) variables in the \code{dframe} data frame.
+#'   names of subpopulation (domain) variables in \code{dframe}.
 #'   If a value is not provided, the value \code{"All_Sites"} is assigned to the
 #'   subpops argument and a factor variable named \code{"All_Sites"} that takes
-#'   the value \code{"All Sites"} is added to the \code{dframe} data frame.  The
+#'   the value \code{"All Sites"} is added to \code{dframe}.  The
 #'   default value is \code{NULL}.
 #'
 #' @param surveyID Character value providing name of the survey ID variable in
-#'   the \code{dframe} data frame.  The default value is \code{"surveyID"}.
+#'   \code{dframe}.  The default value is \code{"surveyID"}.
 #'
 #' @param survey_names Character vector of length two that provides the survey
 #'   names contained in the \code{surveyID} variable in the \code{dframe} data
@@ -77,11 +77,11 @@
 #'   The default is \code{NULL}.
 #'
 #' @param siteID Character value providing name of the site ID variable in
-#'   the \code{dframe} data frame.  For a two-stage sample, the site ID variable
+#'   \code{dframe}.  For a two-stage sample, the site ID variable
 #'   identifies stage two site IDs.  The default value is \code{"siteID"}.
 #'
 #' @param weight Character value providing name of the survey design weight
-#'   variable in the \code{dframe} data frame.  For a two-stage sample, the
+#'   variable in \code{dframe}.  For a two-stage sample, the
 #'   weight variable identifies stage two weights.  The default value is
 #'   \code{"weight"}.
 #'
@@ -94,34 +94,34 @@
 #'   estimate standard error.  The default is \code{FALSE}.
 #'
 #' @param xcoord Character value providing name of the x-coordinate variable in
-#'   the \code{dframe} data frame.  For a two-stage sample, the x-coordinate
+#'   \code{dframe}.  For a two-stage sample, the x-coordinate
 #'   variable identifies stage two x-coordinates.  Note that x-coordinates are
 #'   required for calculation of the local mean variance estimator.  The default
 #'   value is \code{NULL}.
 #'
 #' @param ycoord Character value providing name of the y-coordinate variable in
-#'   the \code{dframe} data frame.  For a two-stage sample, the y-coordinate
+#'   \code{dframe}.  For a two-stage sample, the y-coordinate
 #'   variable identifies stage two y-coordinates.  Note that y-coordinates are
 #'   required for calculation of the local mean variance estimator.  The default
 #'   value is \code{NULL}.
 #'
 #' @param stratumID Character value providing name of the stratum ID variable in
-#'   the \code{dframe} data frame.  The default value is \code{NULL}.
+#'   \code{dframe}.  The default value is \code{NULL}.
 #'
 #' @param clusterID Character value providing the name of the cluster
-#'   (stage one) ID variable in the \code{dframe} data frame.  Note that cluster
+#'   (stage one) ID variable in \code{dframe}.  Note that cluster
 #'   IDs are required for a two-stage sample.  The default value is \code{NULL}.
 #'
 #' @param weight1 Character value providing name of the stage one weight
-#'   variable in the \code{dframe} data frame.  The default value is \code{NULL}.
+#'   variable in \code{dframe}.  The default value is \code{NULL}.
 #'
 #' @param xcoord1 Character value providing the name of the stage one
-#'   x-coordinate variable in the \code{dframe} data frame.  Note that x
+#'   x-coordinate variable in \code{dframe}.  Note that x
 #'   coordinates are required for calculation of the local mean variance
 #'   estimator.  The default value is \code{NULL}.
 #'
 #' @param ycoord1 Character value providing the name of the stage one
-#'   y-coordinate variable in the \code{dframe} data frame.  Note that
+#'   y-coordinate variable in \code{dframe}.  Note that
 #'   y-coordinates are required for calculation of the local mean variance
 #'   estimator.  The default value is \code{NULL}.
 #'
@@ -133,12 +133,12 @@
 #'   arguments \code{weight} and \code{weight1}. The default value is \code{FALSE}.
 #'
 #' @param sweight Character value providing the name of the size weight variable
-#'   in the \code{dframe} data frame.  For a two-stage sample, the size weight
+#'   in \code{dframe}.  For a two-stage sample, the size weight
 #'   variable identifies stage two size weights.  The default value is
 #'   \code{NULL}.
 #'
 #' @param sweight1 Character value providing name of the stage one size weight
-#'   variable in the \code{dframe} data frame.  The default value is \code{NULL}.
+#'   variable in \code{dframe}.  The default value is \code{NULL}.
 #'
 #' @param fpc Object that specifies values required for calculation of the
 #'   finite population correction factor used during variance estimation. The
@@ -211,7 +211,7 @@
 #'   post-stratification and raking by calibrating a sample to the marginal
 #'   totals of variables in a linear regression model. For the \code{calibrate}
 #'   function, the object is a named list, where the names identify factor
-#'   variables in the \code{dframe} data frame.  Each element of the list is a
+#'   variables in \code{dframe}.  Each element of the list is a
 #'   named vector containing the population total for each level of the
 #'   associated factor variable.  For the \code{postStratify} function, the
 #'   object is either a data frame, table, or xtabs object that provides the
@@ -600,7 +600,7 @@ change_analysis <- function(
 
   # If a value was not provided for the subpops (subpopulation names) argument,
   # assign the value "All_Sites" to the subpops argument and create a factor
-  # named "All_Sites" in the dframe data frame that takes the value "All Sites"
+  # named "All_Sites" in dframe that takes the value "All Sites"
 
   if (is.null(subpops)) {
     subpops <- "All_Sites"
@@ -610,7 +610,7 @@ change_analysis <- function(
 
   # If the user wants information for all sites together in addition to the
   # subpops, add the value "All_Sites" to the subpops argument and create a
-  # factor named "All_Sites" in the dframe data frame that takes the value
+  # factor named "All_Sites" in dframe that takes the value
   # "All Sites"
 
   if (!is.null(subpops) && All_Sites) {